+++ packages/leva/src/helpers/monitor.ts
@@ -2,9 +2,6 @@ import { SpecialInputTypes, MonitorInput, MonitorSettings } from '../types/'
 
 const defaultSettings = { graph: false, interval: 100 }
 
-export function monitor(
-  objectOrFn: React.MutableRefObject<any> | Function,
-  settings?: Partial<MonitorSettings>
-): MonitorInput {
+export function monitor(objectOrFn: React.MutableRefObject<any> | Function, settings?: MonitorSettings): MonitorInput {
   return { type: SpecialInputTypes.MONITOR, objectOrFn, settings: { ...defaultSettings, ...settings } }
 }
