+++ packages/leva/src/hooks/useVisiblePaths.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react'
-import { StoreType } from '../store'
 import shallow from 'zustand/shallow'
+import { StoreType } from '../types'
 
 /**
  * Hook used by the root component to get all visible inputs.
